+++ CHANGELOG.md
@@ -2,7 +2,12 @@
 
 ## [4.1.0] - 2021-XX-XX
 
+### ✨ New features
+
 - [#113](https://github.com/estruyf/vscode-front-matter/issues/113): Integrating a local DB for media metadata (description, alt)
+
+### 🎨 Enhancements
+
 - [#117](https://github.com/estruyf/vscode-front-matter/issues/117): Allow to specify a singleline of text in the metadata fields
 
 ## [4.0.1] - 2021-09-24